Started the good, the bad and the ugly - code cleanup.
